Spring Security Done Without RBA

+++ src/main/java/com/sm/journalApp/services/JournalEntryService.java
@@ -9,15 +9,18 @@ import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+
 @RequiredArgsConstructor
 @Component
 @Slf4j
 public class JournalEntryService {
     private final JournalEntryRepository journalEntryRepository;
     private final UserService userService;
+
     @Transactional
     public void saveEntry(JournalEntry journalEntry, String userName) {
         try {
@@ -31,30 +34,40 @@ public class JournalEntryService {
             log.error("Error while processing request", e);
             throw new RuntimeException("An error occurred while saving the entry.", e);
         }
-    } 
+    }
 
 
-    public void saveEntry(JournalEntry journalEntry){
+    public void saveEntry(JournalEntry journalEntry) {
         journalEntryRepository.save(journalEntry);
     }
 
-    public List<JournalEntry> getAll(){
+    public List<JournalEntry> getAll() {
         return journalEntryRepository.findAll();
     }
 
-    public Optional<JournalEntry> getById(ObjectId myId){
-        return  journalEntryRepository.findById(myId);
+    public Optional<JournalEntry> getById(ObjectId myId) {
+        return journalEntryRepository.findById(myId);
     }
 
-    public void deleteById(ObjectId myId, String userName){
-        User user = userService.findByUserName(userName);
+    @Transactional
+    public boolean deleteById(ObjectId myId, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
 //        User user2 = userService.findByUserId(myId);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(myId));
-        userService.saveUser(user);
-        journalEntryRepository.deleteById(myId);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(myId));
+
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(myId);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException(e + " Error while deleting journal entry with id: " + myId);
+        }
+        return removed;
     }
 
-    public List<JournalEntry> findByUserName(String userName){
+    public List<JournalEntry> findByUserName(String userName) {
         User user = userService.findByUserName(userName);
         return user.getJournalEntries();
     }
